requirements.txt
langchain-core
langchain-community
langchain-openai
openai
faiss-cpu
pypdf
tiktoken
python-dotenv
streamlit
